requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
werkzeug==2.3.6
flask-cors==4.0.0
datetime==4.3
geopy==2.3.0
